+++ src/IsNotNumberLiteral.ts
@@ -11,8 +11,7 @@
  */
 
 import { Assert, Equals } from "./common";
-
-type IsNotNumberLiteral<T> = T extends number ? (0 extends T ? (1 extends T ? false : true) : true) : never;
+import { IsNotNumberLiteral } from "./solutions/IsNotNumberLiteral";
 
 type T01 = Assert<Equals<IsNotNumberLiteral<number>, false>>;
 type T02 = Assert<Equals<IsNotNumberLiteral<string>, never>>;
