+++ src/solutions/IsEven.ts
@@ -1,0 +1,1 @@
+export type IsEven<N extends number> = `${N}` extends `${string}${"0" | "2" | "4" | "6" | "8"}` ? true : false;
