+++ src/solutions/IsNever.ts
@@ -1,0 +1,1 @@
+export type IsNever<T> = [T] extends [never] ? true : false;
