+++ src/solutions/MakeExciting.ts
@@ -1,0 +1,1 @@
+export type MakeExciting<S extends string> = `${S}!`;
