+++ src/solutions/SplitString.ts
@@ -1,0 +1,3 @@
+import { UnknownArray } from "../utility";
+
+export type SplitString<S extends string, R extends UnknownArray = []> = S extends `${infer C}${infer M}` ? SplitString<M, [...R, C]> : R;
